Fix nondeterminism due to uninitialized memory. Fixes #35.
The `Patch` class does not initialise various fields, most importantly
not `float m_dscale`, which is used e.g. in numeric optimizations.

It is set in "CpatchOrganizerS::setScales" but that setting uses only
`+=`, `/=` and so on, thus depending on the previous value.
This works "out of luck" in most cases because the memory for patches
are re-used and the final `m_dscale` of the previous patch often
happens to be close to 0.

This change makes my testing invocation of pmvs2 valgrind-clean,
while before it would immediately warn about use of unititialized
values.